Implement inplace recursive quicksort, with its own unit test format.
Made quicksort with Wikipedia's implementation of the in place quicksort.
In order to get the unit test to pass, used the strcmp as the comparison
directly instead of utilizing the function passed in. Will need to figure
out how to do the comparison with the test_cmp function in the unit test.