Implement Hashmap_uniq_set function that adds set insert feature to Hashmaps.
Hashmap_uniq_set checks to see if the same key already exists
within the hash, and does not add to the bucket if it does. Note
that this resolves the ambiguity of having two keys with different
values, as now it always resolves for the older value. Associated
unit tests have also been added to test functionality