Exercise Extra credit finished.
Add 4 extra hash functions, jenkins, hsieh and a broken one. Unit
test update to accomodate additional hash functions, as well as
the move of the jenkins function from hashmap to hashmap_algos.
The bad hash function is "not bad enough", and more work is needed
for it to perform pathelogically bad.